# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LendyPH API",
        "description": "REST API for LendyPH Lending Application",
        "contact": {
            "name": "LendyPH",
            "email": "support@lendyph.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/borrowers/{borrowerId}/co-makers": {
            "get": {
                "tags": [
                    "Co-makers"
                ],
                "summary": "List co-makers for a borrower",
                "operationId": "22e1ad2d34929a809a9190279adc992e",
                "parameters": [
                    {
                        "name": "borrowerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Co-maker list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Borrower not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Co-makers"
                ],
                "summary": "Create co-maker for a borrower",
                "operationId": "850339c750d576cc0eb7f255c64c7760",
                "parameters": [
                    {
                        "name": "borrowerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Maria"
                                    },
                                    "middle_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Santos"
                                    },
                                    "suffix": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "09181234567"
                                    },
                                    "occupation": {
                                        "type": "string"
                                    },
                                    "employer": {
                                        "type": "string"
                                    },
                                    "monthly_income": {
                                        "type": "number",
                                        "example": 20000
                                    },
                                    "relationship_to_borrower": {
                                        "type": "string",
                                        "example": "Spouse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Co-maker created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/co-makers/{id}": {
            "get": {
                "tags": [
                    "Co-makers"
                ],
                "summary": "Show co-maker",
                "operationId": "4830b9aa404310855a56a6c3ea483527",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Co-maker details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Co-makers"
                ],
                "summary": "Update co-maker",
                "operationId": "b4756db11567c1a2c7f4ea685b583de0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "middle_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "suffix": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "contact_number": {
                                        "type": "string"
                                    },
                                    "occupation": {
                                        "type": "string"
                                    },
                                    "employer": {
                                        "type": "string"
                                    },
                                    "monthly_income": {
                                        "type": "number"
                                    },
                                    "relationship_to_borrower": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Co-maker updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Co-makers"
                ],
                "summary": "Delete co-maker",
                "operationId": "24ae93ad04f5da60d4aa188b921fc0bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Co-maker deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}